src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.8=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
bertopic=0.11.0=pypi_0
blas=1.0=openblas
bleach=5.0.1=pypi_0
bokeh=2.4.3=py310hca03da5_0
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.10.11=hca03da5_0
certifi=2022.9.24=py310hca03da5_0
cffi=1.15.1=py310h22df2f2_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=37.0.1=py310h834c97f_0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
datasets=2.4.0=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.5.1=pypi_0
entrypoints=0.4=pypi_0
executing=1.0.0=pypi_0
fastjsonschema=2.16.1=pypi_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.8.0=pypi_0
fonttools=4.37.1=pypi_0
freetype=2.11.0=h1192e45_0
frozenlist=1.3.1=pypi_0
fsspec=2022.8.2=pypi_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
gmp=6.2.1=hc377ac9_3
gnutls=3.6.15=h887c41c_0
hdbscan=0.8.28=pypi_0
huggingface-hub=0.8.1=pypi_0
icu=68.1=hc377ac9_0
idna=3.3=pyhd3eb1b0_0
ipykernel=6.15.2=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=py310hca03da5_0
joblib=1.1.0=pypi_0
jpeg=9e=h1a28f6b_0
jsonschema=4.15.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.3.5=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=4.11.1=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.2=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
libcxx=12.0.0=hf6beb65_1
libffi=3.4.2=hc377ac9_4
libgfortran=5.0.0=11_2_0_he6877d6_26
libgfortran5=11.2.0=he6877d6_26
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
libopenblas=0.3.20=hea475bc_0
libopus=1.3=h1a28f6b_1
libpng=1.6.37=hb8d0fd4_0
libtasn1=4.16.0=h1a28f6b_0
libtiff=4.2.0=h01837e1_1
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.2=h68602c7_0
libwebp-base=1.2.2=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
line-profiler=3.5.1=pypi_0
llvm-openmp=12.0.0=haf9daa7_1
llvmlite=0.39.0=pypi_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=hc377ac9_0
markupsafe=2.1.1=py310h1a28f6b_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mpmath=1.2.1=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
nbclient=0.6.7=pypi_0
nbconvert=7.0.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=h1a28f6b_3
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=h84b5d62_1
networkx=2.8.6=pypi_0
nltk=3.7=pypi_0
nose=1.3.7=pypi_0
notebook=6.4.12=pypi_0
numba=0.56.0=pypi_0
numpy=1.22.4=pypi_0
numpy-base=1.21.5=py310h742c864_3
openh264=1.8.0=h98b2900_0
openssl=1.1.1s=h1a28f6b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py310h4d1bdd5_1
pip=22.2.2=pypi_0
plotly=5.10.0=pypi_0
pprofile=2.1.0=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
psutil=5.9.1=pypi_0
psycopg2=2.9.3=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-heat=0.0.6=pypi_0
py-heat-magic=0.0.2=pypi_0
pyarrow=9.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pymongo=4.2.0=pypi_0
pynndescent=0.5.7=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310hca03da5_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.4=hbdb9e5c_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.10_0
pytz=2022.2.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=23.2.1=pypi_0
qtconsole=5.3.2=pypi_0
qtpy=2.2.0=pypi_0
readline=8.1.2=h1a28f6b_1
regex=2022.7.25=pypi_0
requests=2.28.1=py310hca03da5_0
responses=0.18.0=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=61.2.0=py310hca03da5_0
shifterator=0.3.0=pypi_0
six=1.16.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.40=pypi_0
sqlite=3.39.2=h1058600_0
stack-data=0.5.0=pypi_0
sympy=1.11.1=pypi_0
tenacity=8.0.1=pypi_0
terminado=0.15.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.12.1=pypi_0
torchvision=0.13.1=py310_cpu
tornado=6.2=py310h1a28f6b_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
transformers=4.21.1=pypi_0
tweet-query=0.1.0=dev_0
typing_extensions=4.3.0=py310hca03da5_0
tzdata=2022a=hda174b7_0
ujson=5.4.0=pypi_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.11=py310hca03da5_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.2=pypi_0
x264=1!152.20180806=h1a28f6b_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h1a28f6b_1
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=pypi_0
zlib=1.2.12=h5a0b063_2
zstd=1.5.2=h8574219_0
